Fixed CW interferer to be pure sine wave
If we want to send just a CW (Continuous wave)
signal, then all we need to do is send a constant
amplitude to the USRP.

Then, the USRP upconverts it by modulating it against
a pure sine wave at the tx frequency.

A pure sine wave modulated with a constant results in
a pure sine wave as desired.