Added termination messages between controller and nodes
After setting up the scenario, the controller continually listens
for messages from any of the involved nodes. When it receives a
termination message it sends a similar message to all the other
nodes. When a node receives a termination message it terminates
the process. This can be verified by running the controller in
manual mode and then run the other nodes with different run times.
The node with the shortest run time will cause the entire scenario
to be terminated.