Removed CRTS_AP and renamed CRTS_UE to CRTS_CR
I did this because I think it will be better to create an
all purpose CR than differentiating between UEs and APs.
I'm thinking all of the parameters needed to describe either
of these behaviors can be defined for a general purpose CR.
For instance as in LTE a CR might specify that it uses OFDMA
on the DL but SC-FDMA on the UL. The specific network connections
could then be configured either statically or dynamically. In
the end it might be necessary to switch back to separate
implementations, but for now I don't see the need.