New CE uses spectrum sensing for DSA.
When first run, the CE measures the Channel Power to try
to get an estimate of the noise floor.
Then, it transmits until communication degrades, at which point
it stops transmitting and beings spectrum sensing.
It only resumes transmitting when the channel power is less then
the previously meausred noise floor.