Set setup_env command for using CRTS_CR w/o sudo
Running
$ sudo make setup_env
will now give all users on the system sudo priveledges
to run the exact ip and route commands
that are used by CRTS_CR for setting up and connecting
with the TUN interface.
Further, they will be able to run these commands
without a password.
They won't be given the ability to run anything else.